refactor: Renames and split Xml generation into partials
