+++ src/WarHub.ArmouryModel.Source.CodeGeneration/Generators/SourceVisitorGenericVisitPartialGenerator.cs
@@ -55,7 +55,7 @@ namespace WarHub.ArmouryModel.Source.CodeGeneration
                             ThisExpression()
                             .MemberAccess(
                                 IdentifierName(Names.DefaultVisit))
-                            .InvokeWithArguments(
+                            .Invoke(
                                 IdentifierName(Node))));
             }
         }
